refactor(dashboard-tsr): extract part-log + heatmap pure logic, add tests (#1392)
Split part-log-charts chart primitives into part-log-charts-parts and the
KPI aggregation into part-log-derive; extract the heatmap timezone/wall-clock
+ stats helpers into query-count-heatmap-time. Add bun:test coverage for both
pure modules (13 tests). Component exports unchanged.

Part of epic #1392
